⚡ Use token instead of cookies

+++ src/commands/deleteFolder.py
@@ -12,8 +12,8 @@ def deleteFolder(message, called=False):
         ac = dbSql.getDefaultAc(userId)
 
         #! If user has an account
-        if ac:
-            account = Seedr(cookie=ac['cookie'])
+        if ac and ac['token']:
+            account = Seedr(token=ac['token'])
 
             id = message.data[7:] if called else message.text[8:]
             response = account.deleteFolder(id).json()
